Queue refactor
