+++ packages/app/components/Queue/Queue.tsx
@@ -1,0 +1,585 @@
+import { ReactElement, useCallback, useState } from 'react'
+import { useQueue } from '../../hooks/useQueue'
+import { useQuestions } from '../../hooks/useQuestions'
+import { useProfile } from '../../hooks/useProfile'
+import {
+  ClosedQuestionStatus,
+  ERROR_MESSAGES,
+  LimboQuestionStatus,
+  OpenQuestionStatus,
+  Question,
+  QuestionStatusKeys,
+  Role,
+} from '@koh/common'
+import { useTAInQueueInfo } from '../../hooks/useTAInQueueInfo'
+import { useCourse } from '../../hooks/useCourse'
+import {
+  QueueInfoColumn,
+  QueueInfoColumnButton,
+} from './QueueListSharedComponents'
+import { Card, Col, Popconfirm, Row, Tooltip, notification } from 'antd'
+import TACheckinButton from '../Today/TACheckinButton'
+import styled from 'styled-components'
+import { useStudentQuestion } from '../../hooks/useStudentQuestion'
+import { API } from '@koh/api-client'
+import { useRoleInCourse } from '../../hooks/useRoleInCourse'
+import CantFindModal from './Student/StudentCantFindModal'
+import StudentRemovedFromQueueModal from './Student/StudentRemovedFromQueueModal'
+import StudentQueueCard from './Student/StudentQueueCard'
+import StudentBanner from './Student/StudentBanner'
+import { mutate } from 'swr'
+import QuestionForm from './Student/QuestionForm'
+import { useDraftQuestion } from '../../hooks/useDraftQuestion'
+import { useLocalStorage } from '../../hooks/useLocalStorage'
+import { AddStudentsModal } from './TA/TAAddStudent'
+import { EditQueueModal } from './TA/EditQueueModal'
+import PropTypes from 'prop-types'
+
+const HelpNextButton = styled(QueueInfoColumnButton)`
+  color: white;
+  background: #2a9187;
+  &:hover,
+  &:focus {
+    color: white;
+    background: #39aca1;
+  }
+`
+
+const EditQueueButton = styled(QueueInfoColumnButton)`
+  color: #212934;
+`
+
+const Container = styled.div`
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  margin-top: 10px;
+  @media (min-width: 650px) {
+    margin-top: 0;
+    flex-direction: row;
+    height: 100%;
+  }
+`
+
+const QueueListContainer = styled.div`
+  flex-grow: 1;
+  @media (min-width: 650px) {
+    margin-top: 32px;
+  }
+`
+
+const JoinButton = styled(QueueInfoColumnButton)`
+  background-color: #3684c6;
+  color: white;
+`
+
+const VerticalDivider = styled.div`
+  @media (min-width: 650px) {
+    border-right: 1px solid #cfd6de;
+    margin: 0 16px;
+  }
+  @media (min-width: 1000px) {
+    margin: 0 32px;
+  }
+`
+
+const StudentHeaderCard = styled(Card)`
+  height: 64px;
+  padding-left: 8px;
+  padding-right: 8px;
+  background: inherit;
+`
+
+const HeaderText = styled.div`
+  font-size: 14px;
+  font-weight: 500;
+  line-height: 22px;
+  color: #8895a6;
+  font-variant: small-caps;
+`
+
+const PopConfirmTitle = styled.div`
+  max-width: 400px;
+`
+
+const CenterRow = styled(Row)`
+  align-items: center;
+`
+
+interface QueuePageProps {
+  qid: number
+  cid: number
+}
+
+export default function QueuePage({ qid, cid }: QueuePageProps): ReactElement {
+  const { queue } = useQueue(qid)
+  const { questions, mutateQuestions } = useQuestions(qid)
+  const { isCheckedIn, isHelping } = useTAInQueueInfo(qid)
+  const [queueSettingsModal, setQueueSettingsModal] = useState(false)
+  const [addStudentsModal, setAddStudentsModal] = useState(false)
+  const { studentQuestion, studentQuestionIndex } = useStudentQuestion(qid)
+  const [showJoinPopconfirm, setShowJoinPopconfirm] = useState(false)
+  const profile = useProfile()
+  const { course } = useCourse(cid)
+  const [popupEditQuestion, setPopupEditQuestion] = useState(false)
+  const role = useRoleInCourse(cid)
+  const { deleteDraftQuestion } = useDraftQuestion()
+  const [isJoining, setIsJoining] = useState(
+    questions &&
+      studentQuestion &&
+      studentQuestion?.status !== OpenQuestionStatus.Queued,
+  )
+  const [isFirstQuestion, setIsFirstQuestion] = useLocalStorage(
+    'isFirstQuestion',
+    true,
+  )
+
+  const helpingQuestions = questions?.questionsGettingHelp?.filter(
+    (q) => q.taHelped.id === profile.id,
+  )
+
+  // const hasUnresolvedRephraseAlert = questions?.unresolvedAlerts
+  //     ?.map((payload) => (payload as RephraseQuestionPayload).questionId)
+  //     .includes(selectedQuestionId);
+
+  const nextQuestion =
+    questions?.priorityQueue[0] || // gets the first item of priority queue if it exists
+    questions?.queue?.find(
+      (question) => question.status === QuestionStatusKeys.Queued,
+    )
+
+  const staffCheckedIntoAnotherQueue = course?.queues.some(
+    (q) =>
+      q.id !== qid &&
+      q.staffList.some((staffMember) => staffMember.id === profile?.id),
+  )
+
+  let isStaff = false
+  profile.courses.forEach((course) => {
+    if (
+      course.course.id === cid &&
+      (course.role === Role.PROFESSOR || course.role === Role.TA)
+    ) {
+      isStaff = true
+    }
+  })
+
+  const studentQuestionId = studentQuestion?.id
+  const studentQuestionStatus = studentQuestion?.status
+
+  const leaveQueue = useCallback(async () => {
+    await API.questions.update(studentQuestionId, {
+      status: ClosedQuestionStatus.ConfirmedDeleted,
+    })
+
+    setIsJoining(false)
+    await mutateQuestions()
+  }, [mutateQuestions, studentQuestionId])
+
+  const rejoinQueue = useCallback(async () => {
+    await API.questions.update(studentQuestionId, {
+      status: OpenQuestionStatus.PriorityQueued,
+    })
+    await mutateQuestions()
+  }, [mutateQuestions, studentQuestionId])
+
+  const joinQueueAfterDeletion = useCallback(async () => {
+    await API.questions.update(studentQuestion?.id, {
+      status: ClosedQuestionStatus.ConfirmedDeleted,
+    })
+    await mutateQuestions()
+    const newQuestion = await API.questions.create({
+      text: studentQuestion.text,
+      questionType: studentQuestion?.questionType,
+      queueId: qid,
+      location: studentQuestion?.location,
+      force: true,
+      groupable: false,
+    })
+    await API.questions.update(newQuestion.id, {
+      status: OpenQuestionStatus.Queued,
+    })
+    await mutateQuestions()
+  }, [mutateQuestions, qid, studentQuestion])
+
+  async function onHelpQuestion(questionId: number): Promise<void> {
+    try {
+      await API.questions.update(questionId, {
+        status: OpenQuestionStatus.Helping,
+      })
+    } catch (e) {
+      if (
+        e.response?.status === 401 &&
+        e.response?.data?.message ===
+          ERROR_MESSAGES.questionController.updateQuestion.otherTAHelping
+      ) {
+        notification.open({
+          message: 'Another TA is currently helping the student',
+          description:
+            'This happens when another TA clicks help at the exact same time',
+          type: 'error',
+          duration: 3,
+          className: 'hide-in-percy',
+          style: {
+            width: 450,
+          },
+        })
+      }
+    }
+  }
+
+  const joinQueueOpenModal = useCallback(
+    async (force: boolean) => {
+      try {
+        const createdQuestion = await API.questions.create({
+          queueId: Number(qid),
+          text: '',
+          force: force,
+          questionType: null,
+          groupable: false,
+        })
+        const newQuestionsInQueue = [...questions?.queue, createdQuestion]
+        await mutateQuestions({ ...questions, queue: newQuestionsInQueue })
+        setPopupEditQuestion(true)
+        return true
+      } catch (e) {
+        if (
+          e.response?.data?.message?.includes(
+            ERROR_MESSAGES.questionController.createQuestion.oneQuestionAtATime,
+          )
+        ) {
+          return false
+        }
+        return true
+        // TODO: how should we handle error that happens for another reason?
+      }
+    },
+    [mutateQuestions, qid, questions],
+  )
+
+  const helpNext = async () => {
+    await onHelpQuestion(nextQuestion.id)
+    mutateQuestions()
+    // openTeams();
+  }
+
+  const openEditModal = useCallback(async () => {
+    mutate(`/api/v1/queues/${qid}/questions`)
+    setPopupEditQuestion(true)
+  }, [qid])
+
+  const closeEditModal = useCallback(() => {
+    setPopupEditQuestion(false)
+    setIsJoining(false)
+  }, [])
+
+  const leaveQueueAndClose = useCallback(() => {
+    //delete draft when they leave the queue
+    deleteDraftQuestion()
+    leaveQueue()
+    closeEditModal()
+  }, [deleteDraftQuestion, leaveQueue, closeEditModal])
+
+  const finishQuestion = useCallback(
+    async (
+      text: string,
+      questionType: QuestionType,
+      groupable: boolean,
+      location: string,
+    ) => {
+      const updateStudent = {
+        text,
+        questionType,
+        groupable,
+        status:
+          studentQuestionStatus === OpenQuestionStatus.Drafting
+            ? OpenQuestionStatus.Queued
+            : studentQuestionStatus,
+        location,
+      }
+
+      const updatedQuestionFromStudent = await API.questions.update(
+        studentQuestionId,
+        updateStudent,
+      )
+
+      const newQuestionsInQueue = questions?.queue?.map((question: Question) =>
+        question.id === studentQuestionId
+          ? updatedQuestionFromStudent
+          : question,
+      )
+
+      // questions are the old questions and newQuestionsInQueue are questions that've been added since.
+      mutateQuestions({
+        ...questions,
+        yourQuestion: updatedQuestionFromStudent,
+        queue: newQuestionsInQueue,
+      })
+    },
+    [studentQuestionStatus, studentQuestionId, questions, mutateQuestions],
+  )
+
+  const finishQuestionAndClose = useCallback(
+    (
+      text: string,
+      qt: QuestionType,
+      groupable: true,
+      router: Router,
+      cid: number,
+      location: string,
+    ) => {
+      deleteDraftQuestion()
+      finishQuestion(text, qt, groupable, location)
+      closeEditModal()
+      if (isFirstQuestion) {
+        notification.warn({
+          style: { cursor: 'pointer' },
+          message: 'Enable Notifications',
+          className: 'hide-in-percy',
+          description:
+            "Turn on notifications for when it's almost your turn to get help.",
+          placement: 'bottomRight',
+          duration: 0,
+          onClick: () => {
+            notification.destroy()
+            setIsFirstQuestion(false)
+            router.push(`/settings?cid=${cid}`)
+          },
+        })
+      }
+    },
+    [
+      deleteDraftQuestion,
+      finishQuestion,
+      closeEditModal,
+      isFirstQuestion,
+      setIsFirstQuestion,
+    ],
+  )
+
+  function RenderQueueInfoCol(): ReactElement {
+    return isStaff ? (
+      <QueueInfoColumn
+        queueId={qid}
+        isStaff={true}
+        buttons={
+          <>
+            <EditQueueButton
+              data-cy="editQueue"
+              onClick={() => setQueueSettingsModal(true)}
+            >
+              Edit Queue Details
+            </EditQueueButton>
+            <EditQueueButton
+              data-cy="addStudents"
+              disabled={!isCheckedIn}
+              onClick={() => setAddStudentsModal(true)}
+            >
+              Add Students
+            </EditQueueButton>
+            <Tooltip
+              title={!isCheckedIn && 'You must check in to help students!'}
+            >
+              <HelpNextButton
+                onClick={helpNext}
+                disabled={!isCheckedIn || !nextQuestion || isHelping}
+                data-cy="help-next"
+              >
+                Help Next
+              </HelpNextButton>
+            </Tooltip>
+
+            <div style={{ marginBottom: '12px' }}>
+              <Tooltip
+                title={
+                  queue.isDisabled && 'Cannot check into a disabled queue!'
+                }
+              >
+                <TACheckinButton
+                  courseId={cid}
+                  room={queue?.room}
+                  disabled={
+                    staffCheckedIntoAnotherQueue ||
+                    isHelping ||
+                    (queue.isProfessorQueue && role !== Role.PROFESSOR) ||
+                    queue.isDisabled
+                  }
+                  state={isCheckedIn ? 'CheckedIn' : 'CheckedOut'}
+                  block
+                />
+              </Tooltip>
+            </div>
+          </>
+        }
+      />
+    ) : (
+      <QueueInfoColumn
+        queueId={qid}
+        isStaff={false}
+        buttons={
+          !studentQuestion && (
+            <Popconfirm
+              title={
+                <PopConfirmTitle>
+                  You already have a question in a queue for this course, so
+                  your previous question will be deleted in order to join this
+                  queue. Do you want to continue?
+                </PopConfirmTitle>
+              }
+              onConfirm={() => joinQueueOpenModal(true)}
+              okText="Yes"
+              cancelText="No"
+              disabled
+              visible={showJoinPopconfirm}
+              onVisibleChange={setShowJoinPopconfirm}
+            >
+              <JoinButton
+                type="primary"
+                disabled={!queue?.allowQuestions || queue?.isDisabled}
+                data-cy="join-queue-button"
+                onClick={async () =>
+                  setShowJoinPopconfirm(!(await joinQueueOpenModal(false)))
+                }
+              >
+                Join Queue
+              </JoinButton>
+            </Popconfirm>
+          )
+        }
+      />
+    )
+  }
+  const QueueHeader = styled.h2`
+    font-weight: 500;
+    font-size: 24px;
+    color: #212934;
+    margin-bottom: 0;
+  `
+
+  const NoQuestionsText = styled.div`
+    font-weight: 500;
+    font-size: 24px;
+    color: #212934;
+  `
+  interface QueueProps {
+    questions: Question[]
+  }
+
+  function RenderQueueQuestions({ questions }: QueueProps) {
+    return (
+      <div data-cy="queueQuestions">
+        {questions?.length === 0 ? (
+          <NoQuestionsText>There are no questions in the queue</NoQuestionsText>
+        ) : (
+          <>
+            <QueueHeader>Queue</QueueHeader>
+            <StudentHeaderCard bordered={false}>
+              <CenterRow>
+                <Col flex="0 0 64px">
+                  <HeaderText>#</HeaderText>
+                </Col>
+                <Col flex="1 1">
+                  <HeaderText>question</HeaderText>
+                </Col>
+                <Col flex="0 0 80px">
+                  <HeaderText>wait</HeaderText>
+                </Col>
+              </CenterRow>
+            </StudentHeaderCard>
+          </>
+        )}
+        {questions?.map((question: Question, index: number) => {
+          return (
+            <StudentQueueCard
+              key={question.id}
+              rank={index + 1}
+              question={question}
+              cid={cid}
+              qid={qid}
+              isStaff={isStaff}
+            />
+          )
+        })}
+      </div>
+    )
+  }
+
+  return (
+    <>
+      <Container>
+        <RenderQueueInfoCol />
+        <VerticalDivider />
+        <QueueListContainer>
+          {isStaff && helpingQuestions && helpingQuestions.length > 0 ? (
+            <>
+              <QueueHeader>You are Currently Helping</QueueHeader>
+
+              {helpingQuestions?.map((question: Question, index: number) => {
+                return (
+                  <StudentQueueCard
+                    key={question.id}
+                    rank={index + 1}
+                    question={question}
+                    cid={cid}
+                    qid={qid}
+                    isStaff={isStaff}
+                  />
+                )
+              })}
+              <br></br>
+              <br></br>
+            </>
+          ) : (
+            <>
+              <StudentBanner
+                queueId={qid}
+                editQuestion={openEditModal}
+                leaveQueue={leaveQueue}
+              />
+              <div style={{ marginTop: '40px' }} />
+            </>
+          )}
+          <RenderQueueQuestions questions={questions?.queue} />
+        </QueueListContainer>
+      </Container>
+      <QuestionForm
+        visible={
+          (questions && !studentQuestion && isJoining) ||
+          // && studentQuestion.status !== QuestionStatusKeys.Drafting)
+          popupEditQuestion
+        }
+        question={studentQuestion}
+        leaveQueue={leaveQueueAndClose}
+        finishQuestion={finishQuestionAndClose}
+        position={studentQuestionIndex + 1}
+        cancel={closeEditModal}
+      />
+      <EditQueueModal
+        queueId={qid}
+        visible={queueSettingsModal}
+        onClose={() => setQueueSettingsModal(false)}
+      />
+      <AddStudentsModal
+        queueId={qid}
+        visible={addStudentsModal}
+        onClose={() => setAddStudentsModal(false)}
+      />
+      <CantFindModal
+        visible={studentQuestion?.status === LimboQuestionStatus.CantFind}
+        leaveQueue={leaveQueue}
+        rejoinQueue={rejoinQueue}
+      />
+      <StudentRemovedFromQueueModal
+        question={studentQuestion}
+        leaveQueue={leaveQueue}
+        joinQueue={joinQueueAfterDeletion}
+      />
+    </>
+  )
+}
+
+QueuePage.propTypes = {
+  questions: PropTypes.shape({
+    queue: PropTypes.arrayOf(PropTypes.instanceOf(Question)),
+  }),
+}
