+++ packages/app/components/Queue/Student/StudentQueueCard.tsx
@@ -1,10 +1,11 @@
-import { Question } from "@koh/common";
-import { Card, Col } from "antd";
-import { ReactElement } from "react";
-import styled from "styled-components";
-import { getWaitTime } from "../../../utils/TimeUtil";
-import { CenterRow, Text } from "../QueueCardSharedComponents";
-import { truncate } from "../QueueUtils";
+import { Question } from '@koh/common'
+import { Card, Col } from 'antd'
+import { ReactElement } from 'react'
+import styled from 'styled-components'
+import { getWaitTime } from '../../../utils/TimeUtil'
+import { CenterRow, Text } from '../QueueCardSharedComponents'
+import { truncate } from '../QueueUtils'
+import TAQueueDetailButtons from '../TA/TAQueueDetailButtons'
 
 const HorizontalStudentCard = styled(Card)`
   margin-bottom: 8px;
@@ -14,16 +15,14 @@ const HorizontalStudentCard = styled(Card)`
   padding-left: 8px;
   padding-right: 8px;
   color: #595959;
-`;
-
-const bodyStyle = {
-  backgroundColor: "#F6FFED",
-};
+`
 
 interface StudentQueueCardProps {
-  question: Question;
-  rank: number;
-  highlighted: boolean;
+  question: Question
+  rank: number
+  cid: number
+  qid: number
+  isStaff: boolean
 }
 
 /**
@@ -31,25 +30,41 @@ interface StudentQueueCardProps {
  */
 export default function StudentQueueCard({
   question,
-  rank,
-  highlighted,
+  cid,
+  qid,
+  isStaff,
 }: StudentQueueCardProps): ReactElement {
   return (
-    <HorizontalStudentCard
-      style={highlighted ? bodyStyle : {}}
-      bordered={!highlighted}
-    >
+    <HorizontalStudentCard>
       <CenterRow>
-        <Col flex="0 0 64px">
-          <Text>{rank}</Text>
-        </Col>
         <Col flex="1 1">
           <Text>{truncate(question.text, 150)}</Text>
+          <div
+            style={{
+              backgroundColor: 'lightblue',
+              borderRadius: '10px',
+              padding: '5px 10px',
+              marginTop: '5px',
+              display: 'inline-block',
+            }}
+          >
+            <Text>{question.questionType}</Text>
+          </div>
         </Col>
         <Col flex="0 0 80px">
           <Text>{getWaitTime(question)}</Text>
         </Col>
+        {isStaff && (
+          <Col>
+            <TAQueueDetailButtons
+              courseId={cid}
+              queueId={qid}
+              question={question}
+              hasUnresolvedRephraseAlert={false}
+            />
+          </Col>
+        )}
       </CenterRow>
     </HorizontalStudentCard>
-  );
+  )
 }
