+++ packages/app/pages/course/[cid]/queue/[qid].tsx
@@ -1,16 +1,14 @@
-import { Role } from '@koh/common'
 import Head from 'next/head'
 import { useRouter } from 'next/router'
 import React, { ReactElement, useEffect } from 'react'
 import styled from 'styled-components'
 import { StandardPageContainer } from '../../../../components/common/PageContainer'
 import NavBar from '../../../../components/Nav/NavBar'
-import StudentQueue from '../../../../components/Queue/Student/StudentQueue'
-import TAQueue from '../../../../components/Queue/TA/TAQueue'
 import { useQueue } from '../../../../hooks/useQueue'
 import { useRoleInCourse } from '../../../../hooks/useRoleInCourse'
 import { useChatbotContext } from '../../../../providers/chatbotProvider'
 import { Spin } from 'antd'
+import QueuePage from '../../../../components/Queue/Queue'
 
 const Container = styled.div`
   flex: 1;
@@ -34,11 +32,7 @@ export default function Queue(): ReactElement {
 
   function RenderQueuePage(): ReactElement {
     if (role && queue) {
-      return Role.STUDENT === role ? (
-        <StudentQueue qid={Number(qid)} cid={Number(cid)} />
-      ) : (
-        <TAQueue qid={Number(qid)} courseId={Number(cid)} />
-      )
+      return <QueuePage qid={Number(qid)} cid={Number(cid)} />
     } else {
       return <Spin />
     }
